pyproject.toml
[tool.poetry]
name = "psapp-clone-backend"
version = "0.1.0"
description = ""
authors = ["Samuel Martins <samuel.ubuntu@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
psnawp = "^2.1.0"
fastapi = "^0.115.3"
zope-component = "^6.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
